Add files via upload
In order to detect moving objects in video surveillance scenes, background subtraction techniques are the most used. In this paper, we developed and tested an efficient background subtraction technique in video surveillance based on the fast-independent component analysis (fast-ICA) method. The proposed technique initiated, first, on the use of a developed fast-ICA algorithm in order to estimate the de-mixing matrix and the denoising matrix parameters. Second, the estimated foreground can simply model by multiplying the data matrix with the de-mixing matrix. After that, the data matrix is multiplied by the denoising matrix for removing the noise.